+++ R/zzz.R
@@ -27,7 +27,7 @@
         ".", "%>%", "%$%", "group_by",  "arrange", "desc",
         "filter", "ungroup", "select", "mutate", "inner_join",
         "read_csv", "write_csv", "as_tibble", "ReinforcementLearning", "na.omit",
-        "h2o.deeplearning", "h2o.loadModel", "h2o.predict", "h2o.saveModel", "head",
+        "h2o.deeplearning", "h2o.loadModel", "h2o.predict", "h2o.saveModel", "head", "tail",
         "if_else", "lag", "mutate_all", "mutate_if", "slice", "str_detect",
         "predict", "read_rds", "summarise", "write.csv", "write_rds", "ymd_hms",
         "filter_all", "full_join", "bind_rows",  "as.h2o", "bind_cols", "testdir"