+++ NEWS.md
@@ -9,6 +9,7 @@
 
 * added more datasets (files) to test functionality
 * building a functionality to create separate models for each asset (adding functions with prefix aml_xxx_xxx.R)
+* build more simple functions to collect data, build model, score data to predict
 
 # lazytrade 0.3.4